GSS-269: Refactor / Bugfixes in the CFDP implementation
