+++ yamcs-core/src/main/java/org/yamcs/cfdp/CfdpTransactionId.java
@@ -2,21 +2,18 @@ package org.yamcs.cfdp;
 
 import java.util.Objects;
 
-public class CfdpTransactionId {
+import org.yamcs.filetransfer.FileTransferId;
+
+public class CfdpTransactionId extends FileTransferId {
     private int sequenceNumber;
-    private long initiatorEntity;
 
     public CfdpTransactionId(long entityId, long sequenceNumber) {
-        this.initiatorEntity = entityId;
-        this.sequenceNumber = (int)sequenceNumber;
+        super(entityId, sequenceNumber);
+        this.sequenceNumber = (int) sequenceNumber;
     }
 
     public int getSequenceNumber() {
         return sequenceNumber;
-    }
-
-    public long getInitiatorEntity() {
-        return initiatorEntity;
     }
 
     public boolean equals(Object o) {
@@ -30,17 +27,17 @@ public class CfdpTransactionId {
             return false;
         }
         CfdpTransactionId other = (CfdpTransactionId) o;
-        return sequenceNumber == other.sequenceNumber && initiatorEntity == other.initiatorEntity;
+        return sequenceNumber == other.sequenceNumber && initiatorEntityId == other.initiatorEntityId;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(sequenceNumber, initiatorEntity);
+        return Objects.hash(sequenceNumber, initiatorEntityId);
     }
     
     @Override
     public String toString() {
-        return initiatorEntity+"_"+sequenceNumber;
+        return initiatorEntityId + "_"+ sequenceNumber;
     }
 
 }
