+++ yamcs-core/src/main/java/org/yamcs/cfdp/OngoingCfdpTransfer.java
@@ -290,7 +290,7 @@ public abstract class OngoingCfdpTransfer implements CfdpFileTransfer {
      * @return
      */
     public long getInitiatorId() {
-        return cfdpTransactionId.getInitiatorEntity();
+        return cfdpTransactionId.getInitiatorEntityId();
     }
 
     /**
